Ames/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
{% if cookiecutter.python_interpreter != 'python3' %}
pandas

# backwards compatibility
pathlib2
{% endif %}